Fix potential missed barrier in `synchronize_rcu`
Whenever any online thread reports itself as quiescent (as opposed to
being reported by the grace period initiator), the initiator must
execute a memory barrier after observing `gp_holdouts` as 0 so it
actually pairs with something on the reader side.

That wasn't what the code used to do, though: it performed the barrier
only if an initial read of `gp_holdouts` was nonzero, but readers might
all report themselves as quiescent by the time we get there.