Add build option for store-buffering test
It looks like some of the store buffering cases the asymmetric barriers
are supposed to prevent never actually happen in practice because the
thread-local `read_lock_nesting` and `need_qs` variables reside in the
same cache line.

By forcibly placing them in separate cache lines, we can actually cause
crashes when the asymmetric barriers are missing, which is pretty cool.